Renamed Pollyclient and mvoed it to language package
